feat(feature/feature-security): add api review, aou follow , and api report summary

+++ src/main/java/com/quocchung/cntt1/techcycle_system/model/User.java
@@ -51,7 +51,7 @@ public class User {
 
   @Column(name = "trust_score")
   @Builder.Default
-  private Integer trustScore = 0;
+  private Double trustScore = 0.0;
 
   @Column(name = "banned_reason", length = 500)
   private String bannedReason;
